// src/app/textbox-on-canvas-v4/types.ts
export type AnchorDirection =
  | "top-left"
  | "top"
  | "top-right"
  | "left"
  | "center"
  | "right"
  | "bottom-left"
  | "bottom"
  | "bottom-right"